# Conflicts:
#	tools/event_inject/include/get_device_node.h
#	tools/event_inject/src/get_device_node.cpp

+++ frameworks/util/network/include/stream_buffer.h
@@ -36,7 +36,7 @@ public:
     DISALLOW_MOVE(StreamBuffer);
 
     void Clean();
-    bool SetReadIdx(size_t idx);
+    bool SetReadIdx(int32_t idx);
 
     bool Read(std::string& buf);
     bool Write(const std::string& buf);
@@ -80,8 +80,8 @@ protected:
     int32_t rCount_ = 0;
     int32_t wCount_ = 0;
 
-    size_t rIdx_ = 0;
-    size_t wIdx_ = 0;
+    int32_t rIdx_ = 0;
+    int32_t wIdx_ = 0;
     char szBuff_[MAX_STREAM_BUF_SIZE] = {};
 };
